[merge] r1086,r1088-r1098,r1102 have been merged from trunk.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist-Java/branches/RELENG_1_2@1103 9a6ac7ee-f497-465d-b832-7029f4913c1d

+++ jp.go.aist.rtm.RTC/src/jp/go/aist/rtm/RTC/util/TimeValue.java
@@ -202,7 +202,7 @@ public class TimeValue {
      */
     public double toDouble(){
         normalize();
-        return this.tv_sec + this.tv_usec/ (double)TIMEVALUE_ONE_SECOND_IN_USECS;
+        return (double)this.tv_sec + (double)this.tv_usec/ (double)TIMEVALUE_ONE_SECOND_IN_USECS;
     }
     
     /**
